+++ pkg/entry/vault_test.go
@@ -227,7 +227,6 @@ func TestVaultUserPassAuth(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-
 	t.Cleanup(func() {
 		err = u.DeleteSecretPolicy(ctx, policyName)
 		if err != nil {